TableNG: Filter ui/ux improvements (#102985)
* feat(table-ng): filter ui/ux improvements
