+++ packages/grafana-ui/src/components/Table/TableNG/TableNG.tsx
@@ -365,8 +365,9 @@ export function TableNG(props: TableNGProps) {
           headerCellRefs,
           isCountRowsSet,
           osContext,
+          rows,
           // INFO: sortedRows is for correct row indexing for cell background coloring
-          rows: sortedRows,
+          sortedRows,
           setContextMenuProps,
           setFilter,
           setIsInspecting,
@@ -567,6 +568,7 @@ export function mapFrameToDataGrid({
     isCountRowsSet,
     osContext,
     rows,
+    sortedRows,
     setContextMenuProps,
     setFilter,
     setIsInspecting,
@@ -662,7 +664,7 @@ export function mapFrameToDataGrid({
       fieldOptions.cellOptions.applyToRow
     ) {
       rowBg = (rowIndex: number): CellColors => {
-        const display = field.display!(field.values.get(rows[rowIndex].__index));
+        const display = field.display!(field.values.get(sortedRows[rowIndex].__index));
         const colors = getCellColors(theme, fieldOptions.cellOptions, display);
         return colors;
       };
