+++ packages/grafana-ui/src/components/Table/TableNG/utils.test.ts
@@ -105,6 +105,7 @@ const sortColumnsRef = { current: [] };
 const mockOptions = {
   osContext: null,
   rows: [],
+  sortedRows: [],
   setContextMenuProps: () => {},
   setFilter: () => {},
   setIsInspecting: () => {},
