+++ packages/grafana-ui/src/components/Table/TableNG/utils.ts
@@ -479,6 +479,7 @@ export interface MapFrameToGridOptions extends TableNGProps {
   isCountRowsSet: boolean;
   osContext: OffscreenCanvasRenderingContext2D | null;
   rows: TableRow[];
+  sortedRows: TableRow[];
   setContextMenuProps: (props: { value: string; top?: number; left?: number; mode?: TableCellInspectorMode }) => void;
   setFilter: React.Dispatch<React.SetStateAction<FilterType>>;
   setIsInspecting: (isInspecting: boolean) => void;
